Add is_schema_valid helper function (#834)
* add is_schema_valid helper function

* move get_invalid_schema_message to public API

* update release notes

* update documentation to include functions

* move tests
