+++ docs/source/api_reference.rst
@@ -194,7 +194,9 @@ General Utils
     :toctree: generated
     :nosignatures:
 
+    get_invalid_schema_message
     init_series
+    is_schema_valid
 
 Demo Data
 =========
