+++ woodwork/accessor_utils.py
@@ -120,3 +120,57 @@ def _get_valid_dtype(series_type, logical_type):
         valid_dtype = logical_type.primary_dtype
 
     return valid_dtype
+
+
+def get_invalid_schema_message(dataframe, schema):
+    """Return a message indicating the reason that the provided schema cannot be used to
+    initialize Woodwork on the dataframe. If the schema is valid for the dataframe,
+    None will be returned.
+
+    Args:
+        dataframe (DataFrame): The dataframe against which to check the schema.
+        schema (ww.TableSchema): The schema to use in the validity check.
+
+    Returns:
+        str or None: The reason that the schema is invalid for the dataframe
+    """
+    dataframe_cols = set(dataframe.columns)
+    schema_cols = set(schema.columns.keys())
+
+    df_cols_not_in_schema = dataframe_cols - schema_cols
+    if df_cols_not_in_schema:
+        return f'The following columns in the DataFrame were missing from the typing information: '\
+            f'{df_cols_not_in_schema}'
+    schema_cols_not_in_df = schema_cols - dataframe_cols
+    if schema_cols_not_in_df:
+        return f'The following columns in the typing information were missing from the DataFrame: '\
+            f'{schema_cols_not_in_df}'
+    for name in dataframe.columns:
+        df_dtype = dataframe[name].dtype
+        valid_dtype = _get_valid_dtype(type(dataframe[name]), schema.logical_types[name])
+        if str(df_dtype) != valid_dtype:
+            return f'dtype mismatch for column {name} between DataFrame dtype, '\
+                f'{df_dtype}, and {schema.logical_types[name]} dtype, {valid_dtype}'
+    if schema.index is not None and isinstance(dataframe, pd.DataFrame):
+        # Index validation not performed for Dask/Koalas
+        if not all(dataframe.index == dataframe[schema.index]):
+            return 'Index mismatch between DataFrame and typing information'
+        elif not dataframe[schema.index].is_unique:
+            return 'Index column is not unique'
+
+
+def is_schema_valid(dataframe, schema):
+    """Check if a schema is valid for initializing Woodwork on a dataframe
+
+    Args:
+        dataframe (DataFrame): The dataframe against which to check the schema.
+        schema (ww.TableSchema): The schema to use in the validity check.
+
+    Returns:
+        boolean: Boolean indicating whether the schema is valid for the dataframe
+    """
+
+    invalid_schema_message = get_invalid_schema_message(dataframe, schema)
+    if invalid_schema_message:
+        return False
+    return True
