+++ woodwork/table_accessor.py
@@ -8,6 +8,7 @@ from woodwork.accessor_utils import (
     _get_valid_dtype,
     _is_dataframe,
     _update_column_dtype,
+    get_invalid_schema_message,
     init_series
 )
 from woodwork.exceptions import (
@@ -606,7 +607,7 @@ class WoodworkTableAccessor:
 
                 # Try to initialize Woodwork with the existing schema
                 if _is_dataframe(result):
-                    invalid_schema_message = _get_invalid_schema_message(result, self._schema)
+                    invalid_schema_message = get_invalid_schema_message(result, self._schema)
                     if invalid_schema_message:
                         warnings.warn(TypingInfoMismatchWarning().get_warning_message(attr, invalid_schema_message, 'DataFrame'),
                                       TypingInfoMismatchWarning)
@@ -617,7 +618,8 @@ class WoodworkTableAccessor:
                 else:
                     # Confirm that the schema is still valid on original DataFrame
                     # Important for inplace operations
-                    invalid_schema_message = _get_invalid_schema_message(self._dataframe, self._schema)
+                    invalid_schema_message = get_invalid_schema_message(self._dataframe, self._schema)
+
                     if invalid_schema_message:
                         warnings.warn(TypingInfoMismatchWarning().get_warning_message(attr, invalid_schema_message, 'DataFrame'),
                                       TypingInfoMismatchWarning)
@@ -910,7 +912,7 @@ def _check_logical_types(dataframe_columns, logical_types):
 def _check_schema(dataframe, schema):
     if not isinstance(schema, TableSchema):
         raise TypeError('Provided schema must be a Woodwork.TableSchema object.')
-    invalid_schema_message = _get_invalid_schema_message(dataframe, schema)
+    invalid_schema_message = get_invalid_schema_message(dataframe, schema)
     if invalid_schema_message:
         raise ValueError(f'Woodwork typing information is not valid for this DataFrame: {invalid_schema_message}')
 
@@ -928,32 +930,6 @@ def _make_index(dataframe, index):
         raise TypeError('Cannot make index on a Koalas DataFrame.')
     else:
         dataframe.insert(0, index, range(len(dataframe)))
-
-
-def _get_invalid_schema_message(dataframe, schema):
-    dataframe_cols = set(dataframe.columns)
-    schema_cols = set(schema.columns.keys())
-
-    df_cols_not_in_schema = dataframe_cols - schema_cols
-    if df_cols_not_in_schema:
-        return f'The following columns in the DataFrame were missing from the typing information: '\
-            f'{df_cols_not_in_schema}'
-    schema_cols_not_in_df = schema_cols - dataframe_cols
-    if schema_cols_not_in_df:
-        return f'The following columns in the typing information were missing from the DataFrame: '\
-            f'{schema_cols_not_in_df}'
-    for name in dataframe.columns:
-        df_dtype = dataframe[name].dtype
-        valid_dtype = _get_valid_dtype(type(dataframe[name]), schema.logical_types[name])
-        if str(df_dtype) != valid_dtype:
-            return f'dtype mismatch for column {name} between DataFrame dtype, '\
-                f'{df_dtype}, and {schema.logical_types[name]} dtype, {valid_dtype}'
-    if schema.index is not None and isinstance(dataframe, pd.DataFrame):
-        # Index validation not performed for Dask/Koalas
-        if not all(dataframe.index == dataframe[schema.index]):
-            return 'Index mismatch between DataFrame and typing information'
-        elif not dataframe[schema.index].is_unique:
-            return 'Index column is not unique'
 
 
 def _raise_init_error():
