+++ woodwork/tests/accessor/test_table_accessor.py
@@ -40,8 +40,7 @@ from woodwork.table_accessor import (
     _check_logical_types,
     _check_time_index,
     _check_unique_column_names,
-    _check_use_standard_tags,
-    _get_invalid_schema_message
+    _check_use_standard_tags
 )
 from woodwork.table_schema import TableSchema
 from woodwork.tests.testing_utils import (
@@ -1116,68 +1115,6 @@ def test_accessor_with_falsy_column_names(falsy_names_df):
     renamed_df = schema_df.ww.rename({0: 'col_with_name'})
     assert 0 not in renamed_df.columns
     assert 'col_with_name' in renamed_df.columns
-
-
-def test_get_invalid_schema_message(sample_df):
-    schema_df = sample_df.copy()
-    schema_df.ww.init(name='test_schema', index='id', logical_types={'id': 'Double', 'full_name': 'PersonFullName'})
-    schema = schema_df.ww.schema
-
-    assert _get_invalid_schema_message(schema_df, schema) is None
-    assert (_get_invalid_schema_message(sample_df, schema) ==
-            'dtype mismatch for column id between DataFrame dtype, int64, and Double dtype, float64')
-
-    sampled_df = schema_df.sample(frac=0.3)
-    assert _get_invalid_schema_message(sampled_df, schema) is None
-
-    dropped_df = schema_df.drop('id', axis=1)
-    assert (_get_invalid_schema_message(dropped_df, schema) ==
-            "The following columns in the typing information were missing from the DataFrame: {'id'}")
-
-    renamed_df = schema_df.rename(columns={'id': 'new_col'})
-    assert (_get_invalid_schema_message(renamed_df, schema) ==
-            "The following columns in the DataFrame were missing from the typing information: {'new_col'}")
-
-
-def test_get_invalid_schema_message_dtype_mismatch(sample_df):
-    schema_df = sample_df.copy()
-    schema_df.ww.init(logical_types={'age': 'Categorical'})
-    schema = schema_df.ww.schema
-
-    incorrect_int_dtype_df = schema_df.ww.astype({'id': 'Int64'})
-    incorrect_bool_dtype_df = schema_df.ww.astype({'is_registered': 'Int64'})
-    incorrect_str_dtype_df = schema_df.ww.astype({'full_name': 'object'})  # wont work for koalas
-    incorrect_categorical_dtype_df = schema_df.ww.astype({'age': 'string'})  # wont work for koalas
-
-    assert (_get_invalid_schema_message(incorrect_int_dtype_df, schema) ==
-            'dtype mismatch for column id between DataFrame dtype, Int64, and Integer dtype, int64')
-    assert (_get_invalid_schema_message(incorrect_bool_dtype_df, schema) ==
-            'dtype mismatch for column is_registered between DataFrame dtype, Int64, and BooleanNullable dtype, boolean')
-    # Koalas backup dtypes make these checks not relevant
-    if ks and not isinstance(sample_df, ks.DataFrame):
-        assert (_get_invalid_schema_message(incorrect_str_dtype_df, schema) ==
-                'dtype mismatch for column full_name between DataFrame dtype, object, and NaturalLanguage dtype, string')
-        assert (_get_invalid_schema_message(incorrect_categorical_dtype_df, schema) ==
-                'dtype mismatch for column age between DataFrame dtype, string, and Categorical dtype, category')
-
-
-def test_get_invalid_schema_message_index_checks(sample_df):
-    if not isinstance(sample_df, pd.DataFrame):
-        pytest.xfail('Index validation not performed for Dask or Koalas DataFrames')
-
-    schema_df = sample_df.copy()
-    schema_df.ww.init(name='test_schema', index='id', logical_types={'id': 'Double', 'full_name': 'PersonFullName'})
-    schema = schema_df.ww.schema
-
-    different_underlying_index_df = schema_df.copy()
-    different_underlying_index_df['id'] = pd.Series([9, 8, 7, 6], dtype='float64')
-    assert (_get_invalid_schema_message(different_underlying_index_df, schema) ==
-            "Index mismatch between DataFrame and typing information")
-
-    not_unique_df = schema_df.replace({3: 1})
-    not_unique_df.index = not_unique_df['id']
-    not_unique_df.index.name = None
-    assert _get_invalid_schema_message(not_unique_df, schema) == 'Index column is not unique'
 
 
 def test_dataframe_methods_on_accessor(sample_df):
